Created Components for countdown and the items card
